pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiida-defects-neb"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
keywords = []
authors = []
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "aiida-core",
  "aiida-vasp",
  "aiida-quantumespresso"
]

[project.optional-dependencies]
docs = [
"mkdocs",
  "mkdocs-material"
]
tests = [
  "pytest"
]
pre-commit = [
  "pre-commit"
]

[dependency-groups]
dev = [
  "aiida-defects-neb[docs,tests,pre-commit]",
]

[tool.hatch.version]
path = "src/aiida_defects_neb/__about__.py"

[tool.hatch.envs.default]
installer = 'uv'

[tool.hatch.envs.docs]
features = ["docs"]
scripts.build = "mkdocs build --clean --strict"
scripts.serve = "mkdocs serve --dev-addr localhost:8000"
scripts.deploy = "mkdocs gh-deploy --force"

[tool.hatch.envs.pre-commit]
features = ["pre-commit"]
scripts.install = "pre-commit install"
scripts.run = "pre-commit run {args}"

[tool.hatch.envs.hatch-test]
features = ["tests"]
randomize = false
parallel = false
run = "pytest {args}"

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.9", "3.10", "3.11", "3.12", "3.13"]

[tool.ruff]
lint.ignore = [
    "TRY003",  # https://docs.astral.sh/ruff/rules/raise-vanilla-args/
    "EM101",   # https://docs.astral.sh/ruff/rules/raw-string-in-exception/
    "EM102",   # https://docs.astral.sh/ruff/rules/f-string-in-exception/
    "PLR2004", # https://docs.astral.sh/ruff/rules/magic-value-comparison/
    "FBT002",  # https://docs.astral.sh/ruff/rules/boolean-default-value-positional-argument/
    "TID252",  # https://docs.astral.sh/ruff/rules/relative-imports/
]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["INP001", "S101"]
